requirements.txt
fastapi
uvicorn
httpx
python-dotenv
jinja2
pydantic
